# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Recipez API
  description: |
    REST API for the Recipez recipe management application.

    ## Authentication
    Most endpoints require JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Authorization Scopes
    Endpoints require specific scopes in the JWT token. Regular users have scopes prefixed with resource types
    (e.g., `recipe:create`, `category:read`). System user (yeschef) has additional administrative scopes.

    ## System User APIs
    Endpoints marked with `[System User]` are restricted to the system user (yeschef) and are used for
    authentication flows and user management. These require system-level scopes.
  version: 1.0.0
  contact:
    name: Recipez Support
  license:
    name: Proprietary
servers:
  - url: /
    description: Current server
...
paths:
...
  /api/code/create:
    post:
      tags:
        - Code
      summary: Create verification code
      description: |
        Create a new verification code for authentication. **[System User Only]**
        Code expires in 5 minutes. Maximum 3 sends per 15-minute window.
        Requires `code:create` scope (system user).
      operationId: createCode
      security:
        - bearerAuth: [code:create]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCodeRequest'
      responses:
        '200':
          description: Code created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodeCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/code/read:
    post:
      tags:
        - Code
      summary: Read verification code
      description: |
        Retrieve verification code details. **[System User Only]**
        Requires `code:read` scope (system user).
      operationId: readCode
      security:
        - bearerAuth: [code:read]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadCodeRequest'
      responses:
        '200':
          description: Code retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodeReadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Code does not exist
  /api/code/update:
    post:
      tags:
        - Code
      summary: Update verification code
      description: |
        Update verification code properties (count, attempts, cooldown). **[System User Only]**
        Implements rate limiting and cooldown periods.
        Requires `code:update` scope (system user).
      operationId: updateCode
      security:
        - bearerAuth: [code:update]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCodeRequest'
      responses:
        '200':
          description: Code updated or rate limit message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          description: Too many codes sent, cooldown active
  /api/code/verify:
    post:
      tags:
        - Code
      summary: Verify code
      description: |
        Verify a user-submitted verification code. **[System User Only]**
        Maximum 3 attempts per code. Triggers cooldown on failures.
        Requires `code:verify` scope (system user).
      operationId: verifyCode
      security:
        - bearerAuth: [code:verify]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyCodeRequest'
      responses:
        '200':
          description: Code verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid code, expired, or too many attempts
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          description: Too many incorrect attempts, cooldown active
  /api/code/delete:
    post:
      tags:
        - Code
      summary: Delete verification code
      description: |
        Delete a verification code. **[System User Only]**
        Requires `code:delete` scope (system user).
      operationId: deleteCode
      security:
        - bearerAuth: [code:delete]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteCodeRequest'
      responses:
        '200':
          description: Code deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ==================== USER ENDPOINTS (SYSTEM USER) ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token with scopes. Include in Authorization header as: `Bearer <token>`

        Token contains:
        - `sub`: User subject ID
        - `scope`: Space-separated list of permission scopes
        - `iat`: Issued at timestamp
        - `exp`: Expiration time (12 hours from issue)
        - `iss`: Token issuer
        - `aud`: Token audience
...
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - insufficient permissions or scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        response:
          type: object
          properties:
            error:
              type: string
              description: Error message
...
    MessageResponse:
      type: object
      properties:
        response:
          type: object
          properties:
            message:
              type: string

    # ==================== HEALTH SCHEMAS ====================
...
    CreateCodeRequest:
      type: object
      required:
        - email
        - session_id
      properties:
        email:
          type: string
          format: email
        session_id:
          type: string
          description: Session ID to associate with the code
    CodeCreateResponse:
      type: object
      properties:
        response:
          type: object
          properties:
            code:
              type: string
              pattern: "^[a-zA-Z0-9]{4}-[a-zA-Z0-9]{4}$"
              description: Generated verification code
    ReadCodeRequest:
      type: object
      required:
        - email
        - session_id
      properties:
        email:
          type: string
          format: email
        session_id:
          type: string
    CodeData:
      type: object
      properties:
        code_id:
          type: string
          format: uuid
        code_count:
          type: integer
          description: Number of times code has been sent
        code_value:
          type: string
          description: Hashed code value
        code_issued_at:
          type: string
          format: date-time
          nullable: true
        code_expires_at:
          type: string
          format: date-time
          nullable: true
        code_attempts:
          type: integer
          description: Number of verification attempts
        code_cooldown:
          type: string
          format: date-time
          nullable: true
          description: Cooldown expiration time if rate limited
        code_email:
          type: string
        code_session_id:
          type: string
          format: uuid
    CodeReadResponse:
      type: object
      properties:
        response:
          type: object
          properties:
            code:
              $ref: '#/components/schemas/CodeData'
    UpdateCodeRequest:
      type: object
      required:
        - email
        - code
      properties:
        email:
          type: string
          format: email
        code:
          type: object
          required:
            - code_id
            - code_count
          properties:
            code_id:
              type: string
              format: uuid
            code_count:
              type: integer
            code_issued_at:
              type: string
              format: date-time
              nullable: true
            code_cooldown:
              type: string
              format: date-time
              nullable: true
    VerifyCodeRequest:
      type: object
      required:
        - received_code
        - code
      properties:
        received_code:
          type: string
          description: User-submitted verification code
        code:
          type: object
          required:
            - code_id
            - code_value
            - code_expires_at
            - code_attempts
          properties:
            code_id:
              type: string
              format: uuid
            code_value:
              type: string
              description: Hashed code to compare against
            code_expires_at:
              type: string
              format: date-time
            code_attempts:
              type: integer
            code_cooldown:
              type: string
              format: date-time
              nullable: true
    DeleteCodeRequest:
      type: object
      required:
        - code_id
      properties:
        code_id:
          type: string
          format: uuid